W244: the float-to-byte conversion sweep — 976 invented bytes removed, and the class is bounded
0 MATCHes lost, 0 new over-window, 0 regressions. Wave 243 found that a float result assigned to a byte
field must convert through a signed intermediate — retail emits `cvt.w.s`/`mfc1`/`sb`, while a direct
`(u8)someFloat` makes b210 emit the full unsigned bias-and-fixup expansion — and that this was worth
**1,564 bytes** in one function and **720** in another. This wave checked all ~155 candidate sites across
44 files to find out how widespread it is.

**The answer is that it is bounded, and that is the wave's main result.** The overwhelming majority of
float-to-byte conversions in this codebase are **genuinely unsigned and already correct**. Agents confirmed
roughly 50 sites as unsigned with the retail evidence cited every time — the `0x4f000000` compare, the
bias subtraction, the `OR 0x80000000`, the `andi 0xff` — across `itfPanel`, `bcm_panel`, `bp_dialog`,
`k_unit`, `y_runtime`, `mdlFile`, `sfl_psel`, `br_panel`, `op_menu`, `sfl_ground`, `brp_birth`, `bp_tex`
and `br_root`. **Wave 243's two big wins were exceptions, not the head of a distribution**, and knowing
that stops a future wave from sweeping this rule blindly.

**Where it did apply, it still paid well.** `btlUnit 00285880` — the known hard case — went 1520 -> 1292
with nd1290 -> **1058**, cutting a pre-existing 352-byte overshoot to 124. Its residual is retail's packed
MMI/COP2 interpolation (`vftoi0`/`qmfc2`/`ppach`/`ppacb`), which cannot be written in plain C, and the
agent again declined to reach for inline asm. `fclCombineList 003dc700` shrank 68 bytes where retail offset
0x00b4 is a direct `cvt.w.s`/`mfc1`/`andi` with no bias at all.

**Two related defects turned out to be more common than the one I was hunting**, and both are worth more
attention than the original:

- **A float where retail has no float at all.** In `_h_camp_status 001236a0` and `00123b70` our source
  carried a float alpha and converted it, while retail simply does `andi $a3, $s0, 0xff` on an `s32` fade —
  no conversion of any kind. Recovering the integer source saved **224 and 212 bytes** for nd850 -> **617**
  and nd692 -> **597**.
- **A float mirror of a value that already exists as an integer.** Three `_h_camp_main` functions were
  converting a float copy of a fade or frame counter that retail masks and stores directly from its `s32`
  original; recovering the real integer source saved 84, 76 and 84 bytes.

**My heuristic census over-counted substantially and agents caught it.** A large share of the 155
candidates were plain integer narrowing — casts of `u32`/`s32` helper returns to a byte field — with no
float involved anywhere. P244Camp declined eleven on that basis, P244Btl found most of the `btlTarget`,
`btlFormation` and `btlOrder` candidates were integer, and P244Panel found only seven of a claimed nine
`itfPanel` sites were real. The grep pattern `(u8)` near a float token is far too loose; **the operand
must actually be float-valued.**

One instructive decline: `fclCombineList 003dc2d0` uses **two different forms for the same expression** —
retail does a direct signed `cvt.w.s` for the packed colour and a separate unsigned bias conversion for the
text argument. Reconstructing both honestly grew the object past its window and worsened nd, so it was
reverted whole. The conversion form is chosen per use site, not per value.

+++ src/Event/Fcl/fclCombineList.c
@@ -1310,7 +1310,7 @@ void fclCombineList003dc700(s32 base_x, s32 base_y, s16 alpha, FclOwner* owner,
             definition = DAT_007ce420[fallback_text_id];
             FUN_0040e3c0(0.0f, base_x, row_y, (byte)alpha, 0x25,
                           (definition->field_02.variant_count - 1) * 2 + 1);
-            scaled_color = ((u32)(u8)(DAT_007cadd0 * (float)alpha)) | 0xffffff00;
+            scaled_color = (s32)(DAT_007cadd0 * (float)alpha) | 0xffffff00;
             FUN_003b32d0(0.0f, base_x + 0x76, row_y + 0x7e, scaled_color,
                           (s8)text_styles[selected_style + 2], 1,
                           DAT_007ce4e4[fallback_text_id], 0x10, 0x6e);
